Chase my tail for two hours because Rob complained about lweapons.

+++ src/weapons.h
@@ -98,7 +98,7 @@ public:
     
     weapon(weapon const &other);
     //weapon(fix X,fix Y,fix Z,int Id,int Type,int pow,int Dir, int Parentid, int prntid, bool isDummy=false);
-    weapon(fix X,fix Y,fix Z,int Id,int Type,int pow,int Dir, int Parentid, int prntid, bool isDummy=false, byte script_gen=0, byte isLW=0);
+    weapon(fix X,fix Y,fix Z,int Id,int Type,int pow,int Dir, int Parentid, int prntid, bool isDummy=false, int script_gen=0, int isLW=0);
     virtual ~weapon();
     void LOADGFX(int wpn);
     bool Dead();
